// src/components/sections/RCMAssessment/types.ts
export interface DataType {
  key: string;
  no: string;
  process?: string;
  processDescription?: string;
  processObjective?: string;
  processSeverityLevels?: string;
  activity?: string;
  process2?: string;
  stage?: string;
  functions?: string;
  clientSegment?: string;
  operationalUnit?: string;
  division?: string;
  entity?: string;
  unitDepartment?: string;
  productClass?: string;
  productName?: string;
  riskType?: string;
  riskDescription?: string;
  severityImpact?: string;
  probabilityLikelihood?: string;
  classification?: string;
  riskResponseType?: string;
  controlObjectives?: string;
  controlDefinition?: string;
  controlDescription?: string;
  controlResponsibility?: string;
  keyControl?: string;
  levelResponsibility?: string;
  cosoPrinciple?: string;
  operationalApproach?: string;
  operationalFrequency?: string;
  controlClassification?: string;
  internalAuditTest?: string;
  sampleSize?: string;
  grcAdequacy?: string;
  grcEffectiveness?: string;
  explanation?: string;
  soxControlActivity?: string;
  financialStatementAssertion?: string;
  icfrControl?: string;
  // RCM Assessment specific fields
  designAdequacyScore?: string;
  sustainabilityScore?: string;
  scalabilityScore?: string;
  adequacyScore?: string;
  totalScore?: string;
  scale?: string;
  rating?: string;
  designScore?: string;
  operatingScore?: string;
  effectivenessScore?: string;
  objectiveAchievementScore?: string;
  processTimelinessScore?: string;
  resourceConsumptionScore?: string;
  efficiencyScore?: string;
  // COSO related fields
  integrityEthical?: boolean | string;
  boardOversight?: boolean | string;
  orgStructure?: boolean | string;
  commitmentCompetence?: boolean | string;
  managementPhilosophy?: boolean | string;
  assignmentAuthority?: boolean | string;
  hrPolicies?: boolean | string;
  boardAudit?: boolean | string;
  managementControl?: boolean | string;
  externalInfluences?: boolean | string;
  commitmentInternal?: boolean | string;
  enforcementIntegrity?: boolean | string;
  employeeAwareness?: boolean | string;
  accountability?: boolean | string;
  commitmentTransparency?: boolean | string;
  // Other environment fields
  responsibilityMatrix?: boolean | string;
  segregationDuties?: boolean | string;
  reportingLines?: boolean | string;
  mission?: boolean | string;
  visionValues?: boolean | string;
  goalsObjectives?: boolean | string;
  structuresSystems?: boolean | string;
  policiesProcedures?: boolean | string;
  isActive?: boolean;
}